Move generator for identifiers to its own module
Cut-paste
